admin/backend/requirements.txt
# Admin 后端依赖
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
supabase>=2.3.0
python-multipart>=0.0.6
boto3>=1.34.0
pydantic>=2.5.0
